refactor: moving *SearchRequest interfaces from the api.fetch package to the search.request package.
