+++ clients/java/src/main/java/io/camunda/client/api/search/request/UsersByGroupSearchRequest.java
@@ -13,11 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.client.api.fetch;
+package io.camunda.client.api.search.request;
 
 import io.camunda.client.api.search.filter.GroupUserFilter;
-import io.camunda.client.api.search.request.FinalSearchRequestStep;
-import io.camunda.client.api.search.request.TypedSearchRequest;
 import io.camunda.client.api.search.response.GroupUser;
 import io.camunda.client.api.search.sort.GroupUserSort;
