+++ clients/java/src/main/java/io/camunda/client/impl/search/request/GroupSearchRequestImpl.java
@@ -21,9 +21,9 @@ import static io.camunda.client.api.search.request.SearchRequestBuilders.searchR
 
 import io.camunda.client.api.CamundaFuture;
 import io.camunda.client.api.JsonMapper;
-import io.camunda.client.api.fetch.GroupsSearchRequest;
 import io.camunda.client.api.search.filter.GroupFilter;
 import io.camunda.client.api.search.request.FinalSearchRequestStep;
+import io.camunda.client.api.search.request.GroupsSearchRequest;
 import io.camunda.client.api.search.request.SearchRequestPage;
 import io.camunda.client.api.search.response.Group;
 import io.camunda.client.api.search.response.SearchResponse;
