+++ clients/java/src/main/java/io/camunda/client/impl/search/request/UsersByGroupSearchRequestImpl.java
@@ -20,10 +20,10 @@ import static io.camunda.client.api.search.request.SearchRequestBuilders.searchR
 
 import io.camunda.client.api.CamundaFuture;
 import io.camunda.client.api.JsonMapper;
-import io.camunda.client.api.fetch.UsersByGroupSearchRequest;
 import io.camunda.client.api.search.filter.GroupUserFilter;
 import io.camunda.client.api.search.request.FinalSearchRequestStep;
 import io.camunda.client.api.search.request.SearchRequestPage;
+import io.camunda.client.api.search.request.UsersByGroupSearchRequest;
 import io.camunda.client.api.search.response.GroupUser;
 import io.camunda.client.api.search.response.SearchResponse;
 import io.camunda.client.api.search.sort.GroupUserSort;
