+++ clients/java/src/main/java/io/camunda/client/impl/search/request/UsersSearchRequestImpl.java
@@ -21,10 +21,10 @@ import static io.camunda.client.api.search.request.SearchRequestBuilders.userSor
 
 import io.camunda.client.api.CamundaFuture;
 import io.camunda.client.api.JsonMapper;
-import io.camunda.client.api.fetch.UsersSearchRequest;
 import io.camunda.client.api.search.filter.UserFilter;
 import io.camunda.client.api.search.request.FinalSearchRequestStep;
 import io.camunda.client.api.search.request.SearchRequestPage;
+import io.camunda.client.api.search.request.UsersSearchRequest;
 import io.camunda.client.api.search.response.SearchResponse;
 import io.camunda.client.api.search.response.User;
 import io.camunda.client.api.search.sort.UserSort;
